src/test/java/PlaneTest.java
import org.junit.Before;
import org.junit.Test;
import plane.Plane;
import plane.PlaneType;

import static org.junit.Assert.assertEquals;

public class PlaneTest {

    private Plane plane;

    @Before
    public void before() {
        plane = new Plane(PlaneType.BOEING747, 467, 124000);
    }

    @Test
    public void hasPlaneType() {
        assertEquals(PlaneType.BOEING747, plane.getPlaneType());
    }

    @Test
    public void hasCapacity() {
        assertEquals(467, plane.getCapacity());
    }

    @Test
    public void hasTotalWeight() {
        assertEquals(124000, plane.getTotalWeight(), 0.0);
    }
}
